+++ src/components/Table/IconSortable/IconSortableDesc/index.js
@@ -3,36 +3,21 @@ import * as React from 'react';
 import { css, cx } from 'emotion';
 import { colors } from '../../../../variables';
 import { SVGImage } from '../../../SVGImage';
-import sortUp from '../sortUp.svg';
-import sortDown from '../sortDown.svg';
+import sortUp from '../sortAsc.svg';
 import type { IconSortableStyleProps } from '../index';
 
 const styles = {
   icon: css`
-    display: inline-flex;
-    flex-direction: column;
-    vertical-align: middle;
-  `,
-  baseIcon: css`
-    width: 6px;
-    height: 4px;
-    fill: #000000a6;
-    margin: 1px;
-  `,
-  styledIcon: css`
-    fill: ${colors.intentPrimary};
+    width: 16px;
+    height: 10px;
+    fill: ${colors.dark65};
+    transform: rotateX(180deg);
   `
 };
 
 export const IconSortableDesc = ({ className }: IconSortableStyleProps) => (
-  <span className={cx(styles.icon, className)}>
-    <SVGImage
-      className={cx(styles.baseIcon)}
-      glyph={sortUp}
-    />
-    <SVGImage
-      className={cx(styles.baseIcon, styles.styledIcon)}
-      glyph={sortDown}
-    />
-  </span>
+  <SVGImage
+    className={cx(styles.icon, className)}
+    glyph={sortUp}
+  />
 );